requirements.txt
# Core AI and Agent Framework
google-genai>=0.3.0
google-generativeai>=0.8.0

# Automation & Network Tools
ansible>=2.15.0
ansible-lint>=6.22.0
netmiko>=4.3.0
napalm>=4.1.0

# YAML and Configuration
pyyaml>=6.0.1
jinja2>=3.1.2
ruamel.yaml>=0.18.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
molecule>=6.0.0
molecule-docker>=2.1.0

# Code Quality
pylint>=3.0.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Observability & Logging
structlog>=23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.7.0
click>=8.1.7

# CI/CD Integration
python-gitlab>=4.2.0
PyGithub>=2.1.1

# Security
bandit>=1.7.5
safety>=2.3.5
